Move _server to Server trait.
